+++ config/parser.go
@@ -88,7 +88,7 @@ func (p *Parser) parseLines(lines []string) (err error) {
 			p.opts.logLevel = parseString(val, defLogLevel)
 		case "ENABLE_METRICS":
 			p.opts.metrics = parseBool(val, defMetrics)
-		case "HTTP_LISTEN_ADDR":
+		case "HTTP_LISTEN_ADDR", "WAYBACK_LISTEN_ADDR":
 			p.opts.listenAddr = parseString(val, defListenAddr)
 		case "CHROME_REMOTE_ADDR":
 			p.opts.enabledChromeRemote = hasValue(val, defEnabledChromeRemote)